✨ feat: 添加版本显示、截图及签名配置，优化CI流程
- 新增版本号显示，基于Git标签自动生成
- 添加应用签名配置，支持Release构建与发布
- 在README中插入截图，提升可读性
- 优化CI：移除Android SDK缓存，改用更灵活路径
- 调整构建脚本，从Debug改为Release APK
- 修复颜色值转换和空指针检查等代码问题

+++ app/src/main/java/online/youcd/heartrate/data/ble/FloatingWindowManager.kt
@@ -232,7 +232,7 @@ class FloatingWindowManager @Inject constructor(
             background = GradientDrawable().apply {
                 cornerRadius = 24f * density
                 setColor(0xD9121212.toInt())
-                setStroke(dp(1), 0x14FFFFFF.toInt())
+                setStroke(dp(1), 0x14FFFFFF)
             }
             elevation = dp(20).toFloat()
         }
@@ -343,7 +343,7 @@ class FloatingWindowManager @Inject constructor(
             orientation = LinearLayout.HORIZONTAL
             background = GradientDrawable().apply {
                 cornerRadius = dp(3).toFloat()
-                setColor(0x22FFFFFF.toInt())
+                setColor(0x22FFFFFF)
             }
             layoutParams = LinearLayout.LayoutParams(dp(104), dp(6)).apply {
                 topMargin = dp(8)
